metadata.py: add experimental feature -- scan type
Scan the first megabyte of the video and use FFprobe to determine if
there are interlaced frames; if so, decide that the video is
interlaced.

This is of course a dirty hack and an oversimplification. For intance,
there's no distinction between fully interlaced video and telecined
video. (In fact I know little about telecine, so I don't have to plan
to distinguish it.)