falling back to python 2
An OS X or Python 3.x bug breaks file.read when file is larger than 2
GiB (max signed 32-bit int), resulting in an

  IOError: [Errno 22] Invalid argument

each time. The bug should be a 32-bit/64-bit problem. That leads to
failure of SHA-1 on large files. Obviously I don't want to do manual
buffering (just for a not-my-bug). See

 * http://stackoverflow.com/a/24716464
 * http://stackoverflow.com/q/7928784
 * http://stackoverflow.com/a/7937365