util.py: do not fill terminal width with progress...
bar; leave one character of space instead.

Fixes #14.

Turns out cmd.exe and PowerShell on Windows are pretty terminals, and
once we fill the last character of the window's width, the cursor
immediately move to the next line, and even carriage return won't
help. Therefore, we have to leave one empty character for the
cursor (which isn't even a character).

Also moved \r from the end of _FORMAT_STRING to the beginning. This way
the cursor won't always cover or blink over the first character of the
line.