[WIP] storyboard.py: _draw_thumbnail => _create_thumbnail
WARNING: THIS COMMIT IS A WORK IN PROGRESS, AND storyboard.py AS IN THIS
COMMIT WON'T RUN!

Unlike _draw_text_block, which must draw on a supplied camvas,
thumbnails can stand on their own, so it is unreasonable to require that
a thumbnail be drawn to something. I realized this when I try to write a
helper _assemble_images to assemble the metadata sheet, the bare
storyboard, and the banner, when I realized that the same helper can
also be used to assemble the thumbnails into a storyboard, provided that
I have the thumbnails in the form of images.

Unit tests updated accordingly.