[WIP] storyboard.py: write function _tile_images
WARNING: THIS COMMIT IS A WORK IN PROGRESS, AND storyboard.py AS IN THIS
COMMIT WON'T RUN!

This commit introduces the function _tile_images, which can both be used
to assemble thumbnails into a bare storyboard, and assemble different
sections (metadata sheet, bare storyboard, promotional banner, or a
selection of these) into a complete storyboard.

Also introduces unit tests accordingly.