[WIP] storyboard.py: finish StoryBoard rewrite
WARNING: THIS COMMIT IS A WORK IN PROGRESS, AND storyboard.py AS IN THIS
COMMIT WON'T RUN!

We are almost there! Still need to rewrite frame.py for this to
work. (And I know it's actually working since I've made some hotfixes in
frame.py and saw everything in line.

This commit also includes various bug fixes in other places and test
suite improvements.

[ci skip]