manually add autosummary to module docstrings
This way .. automodule:: will show a list of classes and routines of
the module at the top of the page. Otherwise you need to scroll all the
way down just to get a sense of what's in the module.

Also updated index.rst and storyboard.rst to adjust the TOCs.