fix CLI args handling bug in OptionReader
Previously OptionReader.opt() will return None for a CLI argument even
if the argument is not specified on the command line. This is because
argparse.Namespace.__dict__ includes keys whose values are None. Now we
have excluded those None keys (i.e., possible but not specified
arguments).

Also updated unit tests to catch the bug.